http - Remove shim to support pre 1.6 (#463)
